postspecification 분리
